requirements.txt
dependency-injector
fastapi
uvicorn
pyyaml
sqlalchemy
requests
aiofiles
python-multipart
pydantic
passlib
python-jose
bcrypt
deta